add: reference-based chimera-check
(cherry-picked from commit f0ad1863275b0adb53ae536fb352d4ad838360ab)

+++ scripts/001_external.R
@@ -85,10 +85,14 @@ vsearch_usearch_global <- function(query, ref, threshold, global = TRUE, ncpu = 
 }
 
 vsearch_uchime_ref <- function(query, ref, ncpu = local_cpus()) {
-  tquery <- tempfile("query", fileext = ".fasta")
-  on.exit(unlink(c(tquery), force = TRUE))
-  write_sequence(query, tquery)
-  if (is.character(ref) && length(ref) == 1 && file.exists(ref)) {
+  if (checkmate::test_file_exists(query, "r")) {
+    tquery <- query
+  } else {
+    tquery <- tempfile("query", fileext = ".fasta")
+    on.exit(unlink(c(tquery), force = TRUE))
+    write_sequence(query, tquery)
+  }
+  if (checkmate::test_file_exists(ref, "r")) {
     tref <- ref
   } else {
     tref <- tempfile("ref", fileext = ".fasta")
@@ -99,7 +103,7 @@ vsearch_uchime_ref <- function(query, ref, ncpu = local_cpus()) {
   on.exit(unlink(tchimeras), TRUE)
   vs <- system2(
     find_vsearch(),
-    c(
+    args = c(
       "--uchime_ref", tquery,
       "--db", tref,
       "--chimeras", tchimeras,
@@ -695,7 +699,7 @@ fastx_gz_index <- function(file) {
 #'
 #' @return filename of the output file
 #' @rdname fastx_gz
-fastx_gz_extract <- function(infile, index, i, outfile, renumber = FALSE, append = FALSE) {
+fastx_gz_extract <- function(infile, index, i, outfile, renumber = FALSE, append = FALSE, hash = NULL) {
   checkmate::assert_file_exists(infile, "r")
   checkmate::assert_file_exists(index, "r")
   checkmate::assert_integerish(i, lower = 1)
